moved notifications from emergencyButton
