+++ package.json
@@ -25,6 +25,7 @@
     "expo-font": "~10.0.4",
     "expo-linear-gradient": "~11.0.3",
     "expo-location": "~14.0.1",
+    "expo-notifications": "~0.14.0",
     "expo-random": "~12.1.1",
     "expo-secure-store": "~11.1.0",
     "expo-status-bar": "~1.2.0",
@@ -33,9 +34,7 @@
     "react": "17.0.1",
     "react-dom": "17.0.1",
     "react-native": "0.64.3",
-
     "react-native-countdown-circle-timer": "^3.0.9",
-
     "react-native-geolocation": "^1.0.0",
     "react-native-geolocation-service": "^5.3.0-beta.4",
     "react-native-gesture-handler": "~2.1.0",
@@ -48,13 +47,8 @@
     "react-native-system-notification": "^0.2.1",
     "react-native-vector-icons": "^9.1.0",
     "react-native-web": "0.17.1",
-
-    "react-native-webview": "11.15.0"
-
     "react-native-webview": "11.15.0",
-    "tsc": "^2.0.4",
-    "expo-notifications": "~0.14.0"
-
+    "tsc": "^2.0.4"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
